Refactorizacion metodos
