+++ webapp/src/screens/ranking/RankingFiltered.tsx
@@ -26,6 +26,40 @@ export type ObtenerDatosRanking = (
 
 export type GetMedal = (pos: number) => string;
 
+export type RankingEntryShared = {
+    position: number;
+    username: string;
+    value: number;
+    percentage: string;
+};
+
+export function sortData(data: RankingEntryShared[], sortBy: SortRule): RankingEntryShared[] {
+    const sorted = data.slice().sort((a, b) => {
+        if (sortBy === "percentage") {
+            const diff = Number.parseFloat(b.percentage) - Number.parseFloat(a.percentage);
+            return diff || b.value - a.value;
+        }
+        const diff = b.value - a.value;
+        return diff || Number.parseFloat(b.percentage) - Number.parseFloat(a.percentage);
+    });
+    const result: RankingEntryShared[] = [];
+    for (let i = 0; i < sorted.length; i++) {
+        let position: number;
+        if (i === 0) {
+            position = 1;
+        } else {
+            const prev = result[i - 1];
+            if (sorted[i].value === sorted[i - 1].value && sorted[i].percentage === sorted[i - 1].percentage) {
+                position = prev.position;
+            } else {
+                position = i + 1;
+            }
+        }
+        result.push({ ...sorted[i], position });
+    }
+    return result;
+}
+
 /**
  *  Funcino para obtener la medalla correspondiente a una posición en el ranking. 
  * @param pos
