+++ webapp/src/screens/ranking/RankingGeneral.tsx
@@ -1,16 +1,11 @@
 import { useState, useEffect } from "react";
-import type { GetMedal, ObtenerDatosRanking, SortRule} from "./RankingFiltered";
+import type { GetMedal, ObtenerDatosRanking, SortRule, RankingEntryShared } from "./RankingFiltered";
+import { sortData } from "./RankingFiltered";
 import "./RankingFilterTypes.css";
 
 type FilterKey = "victorias" | "derrotas";
 
-// TIPO DE ENTRADA DEL RANKING. 
-type RankingEntry = {
-    position: number; // posición en el ranking 
-    username: string; // nombre del jugador
-    value: number; // número de victorias, derrotas o abandonos según el filtro
-    percentage: string; // porcentaje de victorias, derrotas o abandonos sobre el total de partidas
-};
+type RankingEntry = RankingEntryShared;
 
 const FILTER_LABELS: Record<FilterKey, string> = {
     victorias:   "Victorias",
@@ -26,33 +21,6 @@ const FILTER_COLORS: Record<FilterKey, string> = {
     victorias: "ranking-info--green",
     derrotas: "ranking-info--red",
 };
-
-function sortData(data: RankingEntry[], sortBy: SortRule): RankingEntry[] {
-    const sorted = data.slice().sort((a, b) => {
-        if (sortBy === "percentage") {
-            const diff = Number.parseFloat(b.percentage) - Number.parseFloat(a.percentage);
-            return diff || b.value - a.value;
-        }
-        const diff = b.value - a.value;
-        return diff || Number.parseFloat(b.percentage) - Number.parseFloat(a.percentage);
-    });
-    const result: RankingEntry[] = [];
-    for (let i = 0; i < sorted.length; i++) {
-        let position: number;
-        if (i === 0) {
-            position = 1;
-        } else {
-            const prev = result[i - 1];
-            if (sorted[i].value === sorted[i - 1].value && sorted[i].percentage === sorted[i - 1].percentage) {
-                position = prev.position;
-            } else {
-                position = i + 1;
-            }
-        }
-        result.push({ ...sorted[i], position });
-    }
-    return result;
-}
 
 export default function RankingGeneral({ username, obtenerDatos, getMedal, sortBy }: Readonly<{ username: string; obtenerDatos: ObtenerDatosRanking; getMedal: GetMedal; sortBy: SortRule }>) {
